Refactor: Util 클래스를 만들어서 dateFormatter로 날짜를 문자열로 바꾸게 하는 기능을 하게함 	이유 : 음료 객체는 날짜를 문자열로 만들 책임 및 역할이 아니다. 	       따라서 전반적으로 쓰이는 곳을 말하는 Util이라는 클래스를 생성해서 	       Util에서 해당 역할을 하도록 하였다.
